+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.19] - 2025-12-08
+
+### Changed
+- Updated license to proprietary
+- Block access to /admin from public internet
+
 ## [0.1.18] - 2025-12-08
 
 ### Changed